requirements.txt
asgiref==3.11.0
bleach==6.3.0
deprecation==2.0.7
Django==5.1.15
django-localized-fields==6.7
django-modeltranslation==0.19.19
django-postgres-extra==2.0.9
django-resized==1.0.3
django-widget-tweaks==1.5.0
importlib_metadata==8.7.0
Markdown==3.10
packaging==25.0
pillow==12.0.0
python-dateutil==2.9.0.post0
shortuuid==1.0.13
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
webencodings==0.5.1
zipp==3.23.0
